First Commit
This repository contains a front end project based on HTML and CSS only. This follows the youtube channel playlist of Do Some Coding but only minor changes are done. This also contains one CSS grid and positions tutorial folders. The project named example portfolio showcases some of the screenshots of the website's front end in the project section.